0-create_array.c

+++ 0x0B-malloc_free/0-create_array.c
@@ -19,7 +19,7 @@ char *create_array(unsigned int size, char c)
 
 	if (p == NULL)
 		return (NULL);
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int)size; i++)
 		p[i] = c;
 
 	return (p);
